// client/src/services/openapi/models/ArticoliView.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

import type { MatricolaBasicGet } from './MatricolaBasicGet';

export type ArticoliView = {
    ID: string;
    humanReadableID: string;
    matricole: Array<MatricolaBasicGet>;
    arcodar2: string;
    ardesart: string | null;
    ardessup: string | null;
    aroperat: string | null;
    arcodfam: string | null;
    matricolaId: string | null;
    key: string;
};