// client/src/services/openapi/models/MatricolaBasicGet.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type MatricolaBasicGet = {
    ID: string;
    amcodice: string;
    amcodmag: string;
    amkeysal: string;
    humanReadableID: string;
    lnk_matfor: string | null;
};